requirements.txt
fasttext-wheel
jieba
pandas
lime
shap
numpy
